Move project description builder module into main app
