+++ front/app/api/project_description_builder/__mocks__/projectDescriptionBuilderLayout.ts
