+++ front/app/components/ProjectDescriptionBuilder/ContentViewer/index.tsx
