+++ front/app/components/ProjectDescriptionBuilder/CraftSections/InfoWithAccordions/index.tsx
