+++ front/app/components/ProjectDescriptionBuilder/ProjectDescriptionBuilderEditModePreview/index.tsx
