+++ front/app/containers/ProjectDescriptionBuilder/FullscreenPreview/FullscreenPreview.test.tsx
