+++ front/app/containers/ProjectsShowPage/shared/header/ProjectHeader.tsx
@@ -14,12 +14,12 @@ import { maxPageWidth } from 'containers/ProjectsShowPage/styles';
 
 import ContentContainer from 'components/ContentContainer';
 import FollowUnfollow from 'components/FollowUnfollow';
-import Outlet from 'components/Outlet';
 import {
   HeaderImage,
   HeaderImageContainer,
 } from 'components/ProjectableHeader';
 import ProjectArchivedIndicator from 'components/ProjectArchivedIndicator';
+import ContentViewer from 'components/ProjectDescriptionBuilder/ContentViewer';
 import Button from 'components/UI/Button';
 
 import { useIntl } from 'utils/cl-intl';
@@ -149,7 +149,7 @@ const ProjectHeader = memo<Props>(({ projectId, className }) => {
           {!projectDescriptionBuilderEnabled && (
             <ProjectInfo projectId={projectId} />
           )}
-          <Outlet id="app.ProjectsShowPage.shared.header.ProjectInfo.projectDescriptionBuilder" />
+          <ContentViewer />
         </ContentContainer>
       </Container>
     );
