+++ front/app/modules/index.ts
@@ -33,9 +33,6 @@ import moderationConfiguration, {
   moderationRouteTypes,
 } from './commercial/moderation';
 import posthogConfiguration from './commercial/posthog_integration';
-import projectDescriptionBuilderConfiguration, {
-  descriptionBuilderRouteTypes,
-} from './commercial/project_description_builder';
 import representativenessConfiguration, {
   representativenessRouteTypes,
 } from './commercial/representativeness';
@@ -49,7 +46,6 @@ export type moduleRouteTypes =
   | analyticsRouteTypes
   | bulkIdeaImportRouteTypes
   | moderationRouteTypes
-  | descriptionBuilderRouteTypes
   | representativenessRouteTypes
   | widgetsRouteTypes;
 
@@ -86,9 +82,6 @@ export default loadModules([
   },
   {
     configuration: ideaAssignmentConfiguration,
-  },
-  {
-    configuration: projectDescriptionBuilderConfiguration,
   },
   {
     configuration: bulkIdeaImportConfiguration,
